Split ProxyGenerator.swift into focused extension files
Goal: bring the file under file_length (1000), the enum body under
type_body_length (350), each function under function_body_length (50)
and cyclomatic_complexity (10). Also clear remaining line_length
violations in the touched files.

Before: 1575-line `public enum ProxyGenerator` with a 1090-line body,
plus several long functions including `swiftTypeInfo(for:)`
(complexity 27), `swiftLiteral(from:)` (12), `promptDefaultJSONValue`
(12), `makeActorSource` (76 lines, complexity 12), and
`makeResourceWrapperLines` (60 lines).

After: 169-line main file with a 159-line enum body, plus 10 sibling
files in `Utilities/SwiftMCPUtilityCore/`:

- ProxyGenerator+ActorSource.swift — makeActorSource split into
  appendTypeDefinitions / appendMetadataSection / appendClientHeader /
  makeClientBody / appendToolMethods (each ≤ 30 lines).
- ProxyGenerator+ToolMethods.swift — tool wrapper emitters.
- ProxyGenerator+ResourceMethods.swift — resource codegen + template
  parsing; makeResourceWrapperLines split into appendStaticResource-
  Method / appendTemplateResourceMethod / appendTemplateURIConstruction
  / sortedResources.
- ProxyGenerator+PromptMethods.swift — prompt codegen.
- ProxyGenerator+Header.swift — header comment + doc + buildReturnTypes.
- ProxyGenerator+Identifiers.swift — pascalCase, swiftIdentifier,
  isValidIdentifier, escapeSwiftString, reservedKeywords.
- ProxyGenerator+TypeMapping.swift — SwiftTypeInfo + swiftTypeInfo(for:)
  decomposed into scalarSwiftTypeInfo / integerSwiftTypeInfo /
  encodedScalarSwiftTypeInfo / arraySwiftTypeInfo. Now complexity 5.
- ProxyGenerator+Defaults.swift — default-value / JSON literal helpers,
  with swiftLiteralForArray / swiftLiteralForObject /
  promptDefaultScalarJSONValue / promptDefaultArrayJSONValue /
  encodedPromptDefaultLiteral extracted.
- OpenAPITypeRegistry.swift — extracted from `private class`;
  knownSwiftMCPType split into knownToolContentType +
  knownResourceType; ensureStruct extracted appendStructProperties.
- OpenAPIProxyLoader.swift — extracted loader + spec types.

Access widening from `private` → `internal` for cross-file helpers:
pascalCase, swiftIdentifier, isValidIdentifier, escapeSwiftString,
reservedKeywords, MethodParameter, TemplateVariable, SwiftTypeInfo,
wrapperDocCommentLines, uniqueMethodName, plus OpenAPITypeRegistry /
OpenAPIProxySpec at file scope.

Generator output preserved byte-identical (verified by snapshot tests).
Build clean. swift test --filter ProxyGenerator (9/9), full suite (415).
